# postgres_api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/moneothings": {
            "post": {
                "description": "get moneothing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moneothings"
                ],
                "summary": "get moneothing",
                "parameters": [
                    {
                        "description": "get a moneothing",
                        "name": "moneothingsearchdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.moneothingsearchdto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.moneothing"
                        }
                    }
                }
            }
        },
...
        "/rawdatas": {
            "post": {
                "description": "get rawdatas by value",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rawdatas"
                ],
                "summary": "Show rawdatas",
                "parameters": [
                    {
                        "description": "Get rawdatas",
                        "name": "valuesearchdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.valuesearchdto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.rawdataviewmodel"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "main.moneothing": {
            "type": "object",
            "properties": {
                "Data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.moneothingrawdata"
                    }
                },
                "displayname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "thingid": {
                    "type": "string"
                },
                "uniqueidentifier": {
                    "type": "string"
                }
            }
        },
        "main.moneothingrawdata": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "moneothing": {
                    "$ref": "#/definitions/main.moneothing"
                },
                "rawdata": {
                    "$ref": "#/definitions/main.rawdata"
                },
                "rawdataid": {
                    "type": "integer"
                },
                "thingid": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "main.moneothingsearchdto": {
            "type": "object",
            "properties": {
                "pagenumber": {
                    "type": "integer"
                },
                "pagesize": {
                    "type": "integer"
                },
                "thingid": {
                    "type": "string"
                },
                "uniqueidentifier": {
                    "type": "string"
                }
            }
        },
...
        "main.rawdata": {
            "type": "object",
            "properties": {
                "Data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.moneothingrawdata"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "main.rawdataviewmodel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "main.valuesearchdto": {
            "type": "object",
            "properties": {
                "pagenumber": {
                    "type": "integer"
                },
                "pagesize": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    }
}